+++ src/components/TownAuth/KakaoMap/styles.js
@@ -1,0 +1,9 @@
+import styled from 'styled-components';
+
+export const MapContainer = styled.div`
+  height: calc(100% - ${({ theme }) => theme.style.header_height});
+  .kakao_map {
+    width: 100%;
+    height: 60%;
+  }
+`;
