+++ src/components/TownAuth/index.js
@@ -1,4 +1,5 @@
-export { SearchModal } from './SearchModal/SearchModal';
+export { SearchSection } from './SearchSection/SearchSection';
 export { SearchBar } from './SearchBar/SearchBar';
 export { NeighborList } from './NeighborList/NeighborList';
 export { SubmitModal } from './SubmitModal/SubmitModal';
+export { KakaoMap } from './KakaoMap/KakaoMap';
